+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - CLI: new `mc sdi-data-registries` command for SDI [#265]
 - CLI: new `mc sdi-data-types` command for SDI [#265]
 - CLI: new `mc sdi-file-upload` command for SDI [#265]
+- CLI: new `mc sdi-search-schemas` command for SDI [#265]
 - CLI: fixed bug in `mc aggregates` command with multiple lines of same aggregate
 - CLI: if the authentication is configured, the information about it is removed from help to reduce verbosity.
 - SDK: signal calculation client [#258]
@@ -31,6 +32,7 @@
 ### Contributions 3.13.0
 
 - Thanks to @jupiterbak for his AnomalyDetection contribution. <3 :heart:
+- Thanks to @jupiterbak for his Open Edge Client contributions. <3 :heart:
 
 ## 3.12.0 - (Frost Vienna) - April 2021